refactor: (#150) SequentialUUID의 위치를 변경한다.
