+++ src/main/java/spring/backend/core/infrastructure/jpa/shared/BaseEntity.java
@@ -3,11 +3,10 @@ package spring.backend.core.infrastructure.jpa.shared;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
-import org.hibernate.annotations.GenericGenerator;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import spring.backend.core.infrastructure.persistence.interceptor.SequentialUUID;
+import spring.backend.core.infrastructure.persistence.SequentialUUID;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
