+++ src/main/java/spring/backend/core/infrastructure/persistence/SequentialUUID.java
@@ -1,7 +1,6 @@
-package spring.backend.core.infrastructure.persistence.interceptor;
+package spring.backend.core.infrastructure.persistence;
 
 import org.hibernate.annotations.IdGeneratorType;
-import spring.backend.core.infrastructure.persistence.SequentialUUIDGenerator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,7 +8,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @IdGeneratorType(SequentialUUIDGenerator.class)
-@Target({ElementType.METHOD, ElementType.FIELD})
+@Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface SequentialUUID {
 }
